# internal/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "开源、自托管轻量级发布平台 Ech0 的 API 文档",
        "title": "Ech0 API 文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:6277",
    "basePath": "/api",
    "paths": {
...
        "/backup/import": {
            "post": {
                "description": "用户上传备份文件，成功后恢复数据",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "系统备份"
                ],
                "summary": "恢复数据备份",
                "parameters": [
                    {
                        "type": "file",
                        "description": "备份文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "导入备份失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code 状态码，非0时表示自定义HTTP业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "Data 响应数据，具体内容因接口而异"
                },
                "msg": {
                    "description": "Msg 返回信息，通常是状态描述",
                    "type": "string"
                }
            }
        },
...
    }
}